+++ marginaleffects/hypothesis.py
@@ -2,6 +2,7 @@ import re
 from itertools import compress
 import numpy as np
 import polars as pl
+from .hypothesis_formula import eval_hypothesis_formula
 
 
 def eval_string_hypothesis(x: pl.DataFrame, hypothesis: str, lab: str) -> pl.DataFrame:
@@ -47,33 +48,20 @@ def eval_string_hypothesis(x: pl.DataFrame, hypothesis: str, lab: str) -> pl.Dat
 
 # function extracts the estimate column from a data frame and sets it to x. If `hypothesis` argument is a numpy array, it feeds it directly to lincome_multiply. If lincome is a string, it checks if the string is valid, and then calls the corresponding function.
 def get_hypothesis(x, hypothesis, by=None):
-    msg = f"Invalid hypothesis argument: {hypothesis}. Valid arguments are: 'reference', 'revreference', 'sequential', 'revsequential', 'pairwise', 'revpairwise', a formula string, a numpy array or a float."
+    msg = "Invalid `hypothesis` argument"
 
+    lab = get_hypothesis_row_labels(x, by=None)
     if hypothesis is None or isinstance(hypothesis, (int, float)):
         return x
     if isinstance(hypothesis, np.ndarray):
         out = lincom_multiply(x, hypothesis)
         lab = [f"H{i + 1}" for i in range(out.shape[0])]
         out = out.with_columns(pl.Series(lab).alias("term"))
-    elif isinstance(hypothesis, str) and re.search("=", hypothesis) is not None:
+    elif isinstance(hypothesis, str) and "~" in hypothesis:
+        # lab = [f"b{i}" for i in range(x.shape[0])]
+        out = eval_hypothesis_formula(x, hypothesis, lab=lab)
+    elif isinstance(hypothesis, str) and "=" in hypothesis:
         out = eval_string_hypothesis(x, hypothesis, lab=hypothesis)
-    elif isinstance(hypothesis, str):
-        if hypothesis == "reference":
-            hypmat = lincom_reference(x, by)
-        elif hypothesis == "revreference":
-            hypmat = lincom_revreference(x, by)
-        elif hypothesis == "sequential":
-            hypmat = lincom_sequential(x, by)
-        elif hypothesis == "revsequential":
-            hypmat = lincom_revsequential(x, by)
-        elif hypothesis == "pairwise":
-            hypmat = lincom_pairwise(x, by)
-        elif hypothesis == "revpairwise":
-            hypmat = lincom_revpairwise(x, by)
-        else:
-            raise ValueError(msg)
-        out = lincom_multiply(x, hypmat.to_numpy())
-        out = out.with_columns(pl.Series(hypmat.columns).alias("term"))
     else:
         raise ValueError(msg)
     return out
@@ -118,110 +106,3 @@ def get_hypothesis_row_labels(x, by=None):
     lab = [f"({label})" if "-" in label else label for label in lab]
 
     return lab
-
-
-def lincom_revreference(x, by):
-    lincom = -1 * np.identity(len(x))
-    lincom[0] = 1
-    lab = get_hypothesis_row_labels(x, by)
-    if len(lab) == 0 or len(set(lab)) != len(lab):
-        lab = [f"Row 1 - Row {i + 1}" for i in range(len(lincom))]
-    else:
-        lab = [f"{lab[0]} - {la}" for la in lab]
-    lincom = pl.DataFrame(lincom, schema=lab)
-    lincom = lincom.select(lab[1:])
-    return lincom
-
-
-def lincom_reference(x, by):
-    lincom = np.identity(len(x))
-    lincom[0, :] = -1
-    lab = get_hypothesis_row_labels(x, by)
-    if len(lab) == 0 or len(set(lab)) != len(lab):
-        lab = [f"Row {i + 1} - Row 1" for i in range(len(lincom))]
-    else:
-        lab = [f"{la} - {lab[0]}" for la in lab]
-    if lincom.shape[1] == 1:
-        lincom = pl.DataFrame(lincom, schema=lab)
-    else:
-        lincom = pl.DataFrame(lincom.T, schema=lab)
-    lincom = lincom.select(lab[1:])
-    return lincom
-
-
-def lincom_revsequential(x, by):
-    lincom = np.zeros((len(x), len(x) - 1))
-    lab = get_hypothesis_row_labels(x, by)
-    if len(lab) == 0 or len(set(lab)) != len(lab):
-        lab = [f"Row {i + 1} - Row {i + 2}" for i in range(lincom.shape[1])]
-    else:
-        lab = [f"{lab[i]} - {lab[i + 1]}" for i in range(lincom.shape[1])]
-    for i in range(lincom.shape[1]):
-        lincom[i : i + 2, i] = [1, -1]
-    if lincom.shape[1] == 1:
-        lincom = pl.DataFrame(lincom, schema=lab)
-    else:
-        lincom = pl.DataFrame(lincom.T, schema=lab)
-    return lincom
-
-
-def lincom_sequential(x, by):
-    lincom = np.zeros((len(x), len(x) - 1))
-    lab = get_hypothesis_row_labels(x, by)
-    if len(lab) == 0 or len(set(lab)) != len(lab):
-        lab = [f"Row {i + 2} - Row {i + 1}" for i in range(lincom.shape[1])]
-    else:
-        lab = [f"{lab[i + 1]} - {lab[i]}" for i in range(lincom.shape[1])]
-    for i in range(lincom.shape[1]):
-        lincom[i : i + 2, i] = [-1, 1]
-    if lincom.shape[1] == 1:
-        lincom = pl.DataFrame(lincom, schema=lab)
-    else:
-        lincom = pl.DataFrame(lincom.T, schema=lab)
-    return lincom
-
-
-def lincom_revpairwise(x, by):
-    lab_row = get_hypothesis_row_labels(x, by)
-    lab_col = []
-    flag = len(lab_row) == 0 or len(set(lab_row)) != len(lab_row)
-    mat = []
-    for i in range(len(x)):
-        for j in range(1, len(x)):
-            if i < j:
-                tmp = np.zeros((len(x), 1))
-                tmp[i] = -1
-                tmp[j] = 1
-                mat.append(tmp)
-                if flag:
-                    lab_col.append(f"Row {j + 1} - Row {i + 1}")
-                else:
-                    lab_col.append(f"{lab_row[j]} - {lab_row[i]}")
-    if len(mat) == 1:
-        lincom = pl.DataFrame(mat[0], schema=lab_col)
-    else:
-        lincom = pl.DataFrame(np.hstack(mat).T, schema=lab_col)
-    return lincom
-
-
-def lincom_pairwise(x, by):
-    lab_row = get_hypothesis_row_labels(x, by)
-    lab_col = []
-    flag = len(lab_row) == 0 or len(set(lab_row)) != len(lab_row)
-    mat = []
-    for i in range(len(x)):
-        for j in range(1, len(x)):
-            if i < j:
-                tmp = np.zeros((len(x), 1))
-                tmp[j] = -1
-                tmp[i] = 1
-                mat.append(tmp)
-                if flag:
-                    lab_col.append(f"Row {i + 1} - Row {j + 1}")
-                else:
-                    lab_col.append(f"{lab_row[i]} - {lab_row[j]}")
-    if len(mat) == 1:
-        lincom = pl.DataFrame(mat[0], schema=lab_col)
-    else:
-        lincom = pl.DataFrame(np.hstack(mat).T, schema=lab_col)
-    return lincom